feeds(twitter): improve support for twitter avatars (#16)
As mentioned in the [twtxt thread](https://twtxt.net/twt/ed6sbiq), certain twitter feeds weren't getting their avatars properly handeld. After investigation, I found that this was due to a `Feed{}` being returned from `ValidateTwitterFeed` that was missing a preliminary `Avatar` field, resulting in a missing `Avatar` field until `feeds` was eventually restarted. I tested it and this change should fix that issue.

Also a minor change was introduced; if a twitter avatar was already downloaded, it won't be downloaded again. Prior behavior was to download it every time the feed was updated; this may have resulted in ratelimiting.

Reviewed-on: https://git.mills.io/yarnsocial/feeds/pulls/16
Co-authored-by: servusdei2018 <servusdei2018@noreply@mills.io>
Co-committed-by: servusdei2018 <servusdei2018@noreply@mills.io>